feat: update dependencies
