+++ src/main/java/ru/application/filmorate/controller/FilmController.java
@@ -1,6 +1,6 @@
 package ru.application.filmorate.controller;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import ru.application.filmorate.model.Film;
 import org.springframework.web.bind.annotation.*;
 import ru.application.filmorate.service.FilmService;
@@ -8,27 +8,24 @@ import ru.application.filmorate.service.FilmService;
 import javax.validation.Valid;
 import java.util.List;
 
-@RequestMapping("/films")
 @RestController
+@RequestMapping("/films")
+@RequiredArgsConstructor
 public class FilmController {
     private final FilmService filmService;
 
-    public FilmController(FilmService filmService) {
-        this.filmService = filmService;
-    }
-
     @PostMapping
     public Film addFilm(@Valid @RequestBody Film filmFromRequest) {
-        return filmService.addFilm(filmFromRequest);
+        return filmService.getFilmStorage().addFilm(filmFromRequest);
     }
 
     @PutMapping
     public Film updateFilm(@Valid @RequestBody Film filmFromRequest) {
-        return filmService.updateFilm(filmFromRequest);
+        return filmService.getFilmStorage().updateFilm(filmFromRequest);
     }
 
     @GetMapping
     public List<Film> getFilms() {
-        return filmService.getFilms();
+        return filmService.getFilmStorage().getFilms();
     }
 }
