+++ src/main/java/ru/application/filmorate/controller/UserController.java
@@ -1,5 +1,6 @@
 package ru.application.filmorate.controller;
 
+import lombok.RequiredArgsConstructor;
 import ru.application.filmorate.model.User;
 import org.springframework.web.bind.annotation.*;
 import ru.application.filmorate.service.UserService;
@@ -7,27 +8,24 @@ import ru.application.filmorate.service.UserService;
 import javax.validation.Valid;
 import java.util.List;
 
-@RequestMapping("/users")
 @RestController
+@RequestMapping("/users")
+@RequiredArgsConstructor
 public class UserController {
     private final UserService userService;
 
-    public UserController(UserService userService) {
-        this.userService = userService;
-    }
-
     @PostMapping
     public User createUser(@Valid @RequestBody User user) {
-        return userService.createUser(user);
+        return userService.getUserStorage().createUser(user);
     }
 
     @PutMapping
     public User updateUser(@Valid @RequestBody User user) {
-        return userService.updateUser(user);
+        return userService.getUserStorage().updateUser(user);
     }
 
     @GetMapping
     public List<User> getUsers() {
-        return userService.getUsers();
+        return userService.getUserStorage().getUsers();
     }
 }
