+++ src/main/java/ru/application/filmorate/storage/UserStorage.java
@@ -5,10 +5,9 @@ import ru.application.filmorate.model.User;
 import java.util.List;
 
 public interface UserStorage {
+    User createUser(User userFromRequest);
 
-    void addAsFriend();
+    User updateUser(User userFromRequest);
 
-    void removeFromFriends();
-
-    List<User> listOfCommonFriends();
+    List<User> getUsers();
 }
